+++ src/platforms/pc/platform.c
@@ -244,6 +244,14 @@ void sleep_ms(uint32_t ms) {
 #endif
 }
 
+void sleep_us(uint32_t us) {
+#ifdef WIN32
+  Sleep(us/1000);
+#else
+  usleep(us);
+#endif
+}
+
 ret_t platform_prepare(void) {
   static bool_t inited = FALSE;
   if (inited) {