+++ docs/changes.md
@@ -1,5 +1,9 @@
 # 最新动态
 
+2024/05/28
+  * event\_source\_manager支持微秒级的sleep。
+  * 增加sleep\_us(目前只支持linux)
+
 2024/05/24
   * 修复awtk-web编译错误。
   * 增加函数tk\_int\_is\_in\_array